Restructure
Otherwise the components folder will get unruly
