+++ packages/js/src/dash/hooks/use-fetch.js
@@ -1,7 +1,6 @@
 import { useCallback, useEffect, useRef, useState } from "@wordpress/element";
 import { debounce, identity } from "lodash";
 import { FETCH_DELAY } from "../../shared-admin/constants";
-import { fetchJson } from "./fetch-json";
 
 /**
  * @typedef {Object} FetchResult
@@ -16,6 +15,24 @@ import { fetchJson } from "./fetch-json";
  * @param {RequestInit} options The request options.
  * @returns {Promise<any|Error>} The promise of a result, or an error.
  */
+
+/**
+ * @param {string|URL} url The URL to fetch from.
+ * @param {RequestInit} options The request options.
+ * @returns {Promise<any|Error>} The promise of a result, or an error.
+ */
+const fetchJson = async( url, options ) => {
+	try {
+		const response = await fetch( url, options );
+		if ( ! response.ok ) {
+			// From the perspective of the results, we want to reject this as an error.
+			throw new Error( "Not ok" );
+		}
+		return response.json();
+	} catch ( error ) {
+		return Promise.reject( error );
+	}
+};
 
 /**
  * @param {any[]} dependencies The dependencies for the fetch.
